Update dependencies and config for latest Quasar
Updates dependencies to the latest Quasar versions, including core, extras, and app-vite.
This also includes updates to typescript and @types/node.

Updates the quasar config to use the new defineConfig import and aligns the node target version to match updated node engine requirement.

Explicitly define the module type in package.json.

Adds a missing type declaration file for register-service-worker.

Fixes a type issue in the custom service worker.

Updates crypto-utils to correctly type the salt when creating the encryption key.

Removes the old tsconfig and adds the new tsconfig.json that extends the quasar config.

Minor change to point to the correct component path and uses the route meta title directly in the template.